Treat as different component files
